:truck: Moved settings slice to common
