+++ src/pages/options/layout/settings/HideTime.tsx
@@ -3,7 +3,7 @@ import { Switch } from "@mui/material";
 
 import SettingItem from "../utils/SettingItem";
 import { useAppDispatach } from "../../redux/store";
-import { setHideTime } from "../../redux/settings/settings-slice";
+import { setHideTime } from "../../../../common/redux/settings/settings-slice";
 
 const HideTime = ({ value }: { value: boolean }) => {
   const dispatch = useAppDispatach();
