+++ src/pages/options/layout/settings/Sort.tsx
@@ -3,7 +3,7 @@ import { FormControl, Select, MenuItem } from "@mui/material";
 
 import SettingItem from "../utils/SettingItem";
 import { useAppDispatach } from "../../redux/store";
-import { setSort } from "../../redux/settings/settings-slice";
+import { setSort } from "../../../../common/redux/settings/settings-slice";
 
 const Sort = ({ value }: { value: "last-visit" | "most-visit" }) => {
   const dispatch = useAppDispatach();
