+++ src/pages/options/layout/settings/Theme.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { FormControl, Select, MenuItem } from "@mui/material";
 
 import { useAppDispatach } from "../../redux/store";
-import { setTheme } from "../../redux/settings/settings-slice";
+import { setTheme } from "../../../../common/redux/settings/settings-slice";
 import SettingItem from "../utils/SettingItem";
 
 const Theme = ({ value }: { value: "default" | "dark" | "light" }) => {
